+++ frontend/src/components/AddonDropdown.js
