+++ frontend/src/components/StaticGraph.js
