+++ frontend/src/components/ThresholdSelector.js
